Most basic volume buttons implementation (only info msgs, no volume change)
This is only the very first attempt of implementing the volume buttons and
far from complete. Despite being an almost 1:1 copy of the shutdown code,
it was helpful typing it myself line by line and deal with the code
analyzer.
Also, I'm learning Vim with its myriad plugins that make it a full-fledged
IDE as I go. Imagine an old, one legged man with a weight on his foot doing
a hurdle race.